青福换电

+++ src/main/java/com/xiliulou/electricity/service/FranchiseeBindElectricityBatteryService.java
@@ -15,5 +15,5 @@ public interface FranchiseeBindElectricityBatteryService {
 
     void insert(FranchiseeBindElectricityBattery franchiseeBindElectricityBattery);
 
-    List<FranchiseeBindElectricityBattery> queryByFranchiseeId(Integer id);
+    List<FranchiseeBindElectricityBattery> queryByFranchiseeId(Long id);
 }
